% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energiantuonti}
\alias{energiantuonti}
\title{Energian tuonti ja vienti alkuperämaittain}
\format{
\subsection{\code{energiantuonti}}{

Kehys jossa 925 riviä and 4 saraketta:
\describe{
\item{Alue}{Maan tai määryhmän nimi}
\item{Suunta}{Tuonti tai vienti}
\item{time}{Vuosineljännes}
\item{value}{Arvo (miljoonaa euroa)}
...
}
}
}
\source{
\url{https://statfin.stat.fi/PxWeb/pxweb/fi/StatFin_Passiivi/StatFin_Passiivi__ene__ehk/statfinpas_ehk_pxt_004_202100_fi.px/}
}
\usage{
energiantuonti
}
\description{
Otos energian tuonnista ja viennistä alkuperämaittain
}
\keyword{datasets}
