% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.roboplotr}
\alias{print.roboplotr}
\title{Print method for roboplotr objects}
\usage{
\method{print}{roboplotr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{roboplotr}. This object can be any type that
also includes the class \code{roboplotr}.}

\item{...}{Additional arguments passed to the \code{print} method.}
}
\value{
The function does not return a value. It is used for its side effect
of printing.
}
\description{
This is the print method for objects of class \code{roboplotr}. It temporarily
removes the \code{roboplotr} class attribute and related classes before calling
the next method in the method chain, ensuring that the default print method is
used.
}
