% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{roboplot}
\alias{roboplot}
\title{Creates a plotly object with the desired visual specifications.}
\usage{
roboplot(
  d,
  color,
  title,
  subtitle = "",
  caption,
  legend_orientation = "auto",
  legend_position = "bottom",
  margin = NA,
  trace_color = NULL,
  highlight = NULL,
  zeroline = F,
  rangeslider = FALSE,
  linetype,
  line_width = getOption("roboplot.linewidth"),
  hovertext,
  axis_limits = list(x = c(NA, NA), y = c(NA, NA)),
  plot_type = "scatter",
  plot_mode = "dodge",
  plot_yaxis,
  height = getOption("roboplot.height"),
  facet_split,
  pie_rotation = F,
  legend_maxwidth = NULL
)
}
\arguments{
\item{d}{Data for plotting (Tibble).}

\item{color}{Mandatory. Tibble column from tibble d to use for color (Unquoted string).}

\item{title, subtitle, caption}{labeling for plot}

\item{legend_orientation, legend_position}{Legend positioning.}

\item{margin}{Plot margins if calculated margins do not work (List).}

\item{trace_color}{Trace color for all traces. Either a character string or a character vector of name-value pairs interpretable as colors (Character vector).}

\item{highlight}{Either NULL, a double that limits, or a function that returns a logical determining if a given trace is included in legend and assigned a color.}

\item{zeroline}{Determines zeroline inclusion (Logical or double).}

\item{rangeslider}{Determines rangeslider inclusion (Logical).}

\item{linetype}{Optional. Tibble column from tibble d to use for linetype (Unquoted string).}

\item{line_width}{Line width for line plots. Either a double or a double vector of name-value pairs (Double vector).}

\item{hovertext}{A list describing hovertext items "list(rounding = 1, unit = "\%", extra = "(ennuste)")".}

\item{axis_limits}{Determines the limits of the axes (list(x = c(NA,NA), y = c(NA,NA)))".}

\item{plot_type}{Determines the trace type for either the whole plot, or for all variables defined by color as name-value pairs (Character vector).}

\item{plot_mode}{Determines the barmode for a barplot. Disregarded for scatterplots. (Character).}

\item{plot_yaxis}{Determines the yaxis for a horizontal barplot. Disregarded otherwise. (Unquoted string).}

\item{height}{Height of the plot.}

\item{facet_split}{The column from tibble d to use for facet split (Unquoted string).}

\item{pie_rotation}{Will pie charts be rotated to center the 0° point on middle of the first item of the color variable as factor (Logical).}

\item{legend_maxwidth}{Controls if legend items (and y-axis values for horizontal barplots  longer than this will be collapsed with an ellipsis (Double).}
}
\value{
plotly object
}
\description{
Outputs a plotly object.
}
\examples{
\dontrun{
d <- data("StatFin/kan/ntp/statfin_ntp_pxt_132h.px") |>
dplyr::filter(stringr::str_detect(taloustoimi, "B1GMH|P3KS14_S15"),
tiedot == "Kausitasoitettu ja työpäiväkorjattu sarja, viitevuosi 2015, miljoonaa euroa",
lubridate::year(time) >= 2010) |>
dplyr::group_by(taloustoimi) |>
dplyr::mutate(value = ((value/ lag(value, 4)) -1) * 100) |>
dplyr::ungroup() |>
dplyr::mutate(tiedot = dplyr::case_when(stringr::str_detect(taloustoimi, "B1GMH") ~ "BKT",
TRUE ~ "Yksityinen kulutus") |> forcats::as_factor()
p <- d |> roboplot(tiedot, "BKT ja kulutus", subtitle = "Vuosimuutos",
caption =  "Lähde: Tilastokeskus ja Etu",lwd = F,rangeslider = T,height = 600)
p
}
}
