% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roboplot_continuous_pattern}
\alias{roboplot_continuous_pattern}
\title{\code{\link[=roboplot]{roboplot()}} attempts to bind patterns for continuous line for certain plots}
\usage{
roboplot_continuous_pattern(d, along, pattern, series)
}
\description{
\code{\link[=roboplot]{roboplot()}} attempts to bind patterns for continuous line for certain plots
}
