% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{roboplot_create_widget}
\alias{roboplot_create_widget}
\title{roboplot_create_widget}
\usage{
roboplot_create_widget(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=create_widget]{create_widget()}}. See
\code{?create_widget} for more details.}
}
\value{
A list of classes "plotly" and "html"
}
\description{
File creation from \code{\link[=roboplot]{roboplot()}}s (deprecated)
}
\details{
Write html and and other files from \code{\link[=roboplot]{roboplot()}} plots
}
