% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hovertemplate.R
\name{roboplot_get_hovertemplate_specs}
\alias{roboplot_get_hovertemplate_specs}
\title{Creates a list used to construct roboplot hovertemplate}
\usage{
roboplot_get_hovertemplate_specs(
  frequency = NULL,
  rounding = 1,
  unit = "",
  extra = ""
)
}
\arguments{
\item{frequency}{Character. Determines how hovertemplate dates are formatted.
One of "Annual", "Quarterly", "Monthly", "Weekly", "Daily" or NULL. Default NULL.
With NULL roboplot will try to determine the format based on arg 'd' or roboplot.}

\item{rounding}{Double. Determines the number of small digits of hovertemplate values. Default 1.}

\item{unit}{Character. Unit displayed for hovertemplate values.}

\item{extra}{Character vector. Extra text displayed on hovertemplate.}
}
\value{
A list
}
\description{
Creates a list used to construct roboplot hovertemplate
}
