% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roboplot_make_css}
\alias{roboplot_make_css}
\title{Create a css file or string}
\usage{
roboplot_make_css(css_defs, file = NULL)
}
\arguments{
\item{css_defs}{css style definitions. Each object you provide must be a list of three elements.
The first element will be a vector of the selectors to be styled (e.g. table, th, an id or html
class). If the first element is a vector of length greater than one then the selectors will be
comma separated in the css. The second element will be a vector of the css definitions and the
third element will a vector of the values of those definitions.}

\item{file}{Character sting. If a file name is provided then the css code will be printed into
that file. If the argument is NULL (default) then a string will be returned.}
}
\value{
css definitions.
}
\description{
Create a css file or string
}
\examples{
\dontrun{
roboplot_make_css(list(list('table', c('text-align', 'font-size'), c('center', '20px')),
         list('th', c('background-color', 'height'), c('lightgreen', '30px'))))

roboplot_make_css(list(list(c('table', 'td'), c('text-align', 'font-size'), c('center', '20px')),
         list('th', c('background-color', 'height'), c('lightgreen', '30px'))))

roboplot_make_css(list(list('tr:hover', c('text-align', 'font-size'), c('center', '20px')),
         list('th', c('background-color', 'height'), c('lightgreen', '30px'))))
}

}
