% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_ticks.R
\name{roboplot_set_axes}
\alias{roboplot_set_axes}
\title{Define axis and axis tick formats for \code{\link[=roboplot]{roboplot()}} plots.}
\usage{
roboplot_set_axes(
  x = "time",
  y = "value",
  xticktype = "date",
  yticktype = "numeric"
)
}
\arguments{
\item{x, y}{Character. The name of the column from parameter 'd' of \code{\link[=roboplot]{roboplot()}} used for the axis. Defaults to "time" and "value".}

\item{xticktype, yticktype}{Character. Determines axis formatting within \code{\link[=roboplot]{roboplot()}}. One of "character", "date" or "numeric". Defaults to "date" and "numeric".}
}
\value{
A list.
}
\description{
Define axis and axis tick formats for \code{\link[=roboplot]{roboplot()}} plots.
}
\examples{
# Used to define how axis ticks are formatted inside roboplotr::roboplot()
# if the default "time" with tick formatting as "date" for x-axis and /
# or"value" with tick formatting as "numeric" for y-axis are not suitable.

}
