% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{roboplot_set_caption}
\alias{roboplot_set_caption}
\title{Outputs a string used for roboplot captions.}
\usage{
roboplot_set_caption(
  text = NULL,
  prefix = getOption("roboplot.caption")$prefix,
  updated = getOption("roboplot.caption")$updated,
  .data = NULL,
  line.end = getOption("roboplot.caption")$lineend,
  prepend = NULL,
  append = NULL
)
}
\arguments{
\item{text}{Character. The text for plot caption, probably data source name for argument 'd'.}

\item{prefix}{Character. Prefix inserted before the argument 'text', automatically separated by ": ".}

\item{updated}{Date or logical. If Date, will be inserted as the last update time in the caption. If TRUE, will try to get this information from argument '.data'.}

\item{.data}{Data frame. The data frame the caption attempts to read the time of the last update from.}

\item{line.end}{Character. Character inserted to end of line. Default from roboplot.options.}

\item{append, prepend}{Character vectors. Characters inserted after or before other caption text. Every item in vector will get a new line.}
}
\value{
A string.
}
\description{
Outputs a string used for roboplot captions.
}
