% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{roboplot_set_font}
\alias{roboplot_set_font}
\title{Get a list used as font specifications in \code{\link[=roboplot]{roboplot()}}}
\usage{
roboplot_set_font(
  path = "sans-serif",
  size = 12,
  color = "black",
  bold_title = T,
  type = NULL
)
}
\arguments{
\item{path}{Character. Either "serif", "sans-serif" or a path to .otf or .ttf file.}

\item{size}{Double. Determines font size for the font.}

\item{color}{Character. Must be a hexadecimal color or a valid css color.}

\item{bold_title}{Logical. Only used for title font. Determines if the title is bolded.}

\item{type}{Character. One of "main", "title" or "caption".}
}
\value{
A list
}
\description{
Get a list used as font specifications in \code{\link[=roboplot]{roboplot()}}
}
\examples{
# Used to set fonts used by roboplotr::roboplot(). Only supposed to be called
# inside roboplotr::roboplot_set_options(). 'path' can be a file path or
# string "serif" or "sans-serif", if no specific font path will be provided.
# This will change fonts according to the list output of
# roboplotr::roboplot_set_font(). You can designate font color, size, and
# path, and (for title font only) whether it is bolded. You must also provide
# one of "title", "main" or "caption" as 'type' if an actual filepath is
# provided for 'path'.

roboplot_set_options(
  font_title = roboplot_set_font(size = 17, path = "serif", bold = TRUE),
  font_main =  roboplot_set_font(
    size = 11,
    path = system.file("www","fonts","bladerunner.TTF", package = "roboplotr"),
    color = "darkred",
    type = "main"),
  font_caption = roboplot_set_font(color = "green")
)

d <- energiantuonti |> dplyr::filter(Alue == "Kanada",Suunta == "Tuonti")

d |> roboplot(Alue,"Energian tuonti Kanadasta","Milj. €","Tilastokeskus")

# Note that plotly's own downloadImage does not support external fonts.

if(interactive()) {

  d |>
  roboplot(Alue,"Energian tuonti Kanadasta","Milj. €","Tilastokeskus") |>
  roboplot_create_widget(filepath = tempdir(), artefacts = "img_w")

  utils::browseURL(paste0(tempdir(),"/energian_tuonti_kanadasta_levea.png"))
}

# revert to defaults:
roboplot_set_options(reset = TRUE)
}
