% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modebar.R
\name{roboplot_set_imgdl_specs}
\alias{roboplot_set_imgdl_specs}
\title{Get a list used for \code{\link[=roboplot]{roboplot()}} relayouts for downloaded image files when the appropriate modebar button is pressed}
\usage{
roboplot_set_imgdl_specs(
  x = 1280,
  y = 720,
  mainfont = getOption("roboplot.font.main")$size,
  titlefont = getOption("roboplot.font.title")$size,
  captionfont = getOption("roboplot.font.caption")$size,
  suffix = "_img",
  format = "png"
)
}
\arguments{
\item{x, y}{Double. The dimensions for the image file in pixels the modebar button will produce.}

\item{mainfont, titlefont, captionfont}{Double. The font sizes used in the image file the modebar button will produce.}

\item{suffix}{Character. Suffix attached after the name of the downloaded image file.}

\item{format}{Character. One of "png", "svg", "jpg", or "webp". Define the file format of the downloaded image file.}
}
\value{
A list
}
\description{
Get a list used for \code{\link[=roboplot]{roboplot()}} relayouts for downloaded image files when the appropriate modebar button is pressed
}
\examples{
# Used inside roboplotr::roboplot_set_options() to control image download
# specifications. Parameters 'x' and 'y' control the image dimensions. The
# image uses the font specifications of the main plot, but you might need to
# alter the font sizes with 'mainfont', 'titlefont' and 'captionfont'. You
# can give a suffix for image download filenames (the filename is derived
# from the plot title), and provide any one of "png","svg","jpeg", or "webp"
# for the file format.

roboplot_set_options(
  imgdl_wide =
    roboplot_set_imgdl_specs(
      x = 1400,
      y = 350,
      mainfont = 16,
      titlefont = 24,
      captionfont = 14,
      suffix = "",
      format = "svg")
)

if(interactive()) {

  p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")
  p |> roboplot_create_widget(filepath = tempdir(),
                              artefacts = c("img_w")
  )
  utils::browseURL(paste0(tempdir(),"/energian_tuonti.svg"))
}

# The plotly image downloads are controlled by the modebar, so if you use
# other than 'img_wide' you need to use roboplotr::roboplot_set_options to
# control the modebar accordingly in addition to giving the downloaded image
# specifications (roboplotr::roboplot() does have some defaults in place
# for every size, so you can skip image specifications even if you do add
# use other image download buttons in the modebar). With multiple image
# download buttons the 'suffix' parameter is especially handy.

roboplot_set_options(
  imgdl_wide =
    roboplot_set_imgdl_specs(
      x = 1400,
      y = 350,
      mainfont = 16,
      titlefont = 24,
      captionfont = 14,
      suffix = "_wide"),
  imgdl_narrow =
    roboplot_set_imgdl_specs(
      x = 400,
      y = 600,
      mainfont = 16,
      titlefont = 24,
      captionfont = 14,
      suffix = "_narrow"),
  modebar = c("img_w","img_n","img_s")
)

if(interactive()) {

  p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")
  p |> roboplot_create_widget(filepath = tempdir(),
                              artefacts = c("img_w","img_n","img_s")
  )
  utils::browseURL(paste0(tempdir(),"/energian_tuonti_narrow.png"))
  utils::browseURL(paste0(tempdir(),"/energian_tuonti_pieni.png"))
  utils::browseURL(paste0(tempdir(),"/energian_tuonti_wide.png"))
}

# Revert to defaults:
roboplot_set_options(reset = TRUE)
}
