% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roboplot_string2filename}
\alias{roboplot_string2filename}
\title{Strips a string to filename format for roboplots}
\usage{
roboplot_string2filename(string)
}
\arguments{
\item{string}{The string that will be trunctated to filename}
}
\description{
Strips a string to filename format for roboplots
}
