% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{roboplot_upload_widgets}
\alias{roboplot_upload_widgets}
\title{Uploads the html elements and dependencies to cloud storage. DO NOT USE! WORK IN PROGRESS}
\usage{
roboplot_upload_widgets(files_path, upload_path, overwrite = FALSE)
}
\arguments{
\item{files_path}{The folder where the artefacts to be uploaded are located.}

\item{upload_path}{The gcs folder where the artefacts will be uploaded to.}

\item{overwrite}{If named files exist in the cloud storage, will they be overwritten.}
}
\description{
Uploads the html elements and dependencies to cloud storage. DO NOT USE! WORK IN PROGRESS
}
