% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{roboplotr_automate_png}
\alias{roboplotr_automate_png}
\title{Uses a headless browser to render the png files.}
\usage{
roboplotr_automate_png(p, artefacts, dl_path = getwd())
}
\arguments{
\item{p}{A plotly object.}

\item{artefacts}{A character vector of s(mall), n(arrow), and/or w(ide) corresponding to the expected .png sizes.}

\item{dl_path}{The path where the .png files will be downloaded to. Default is current working directory.}
}
\value{
The plotly object p.
}
\description{
Uses a headless browser to render the png files.
}
