% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_patterns.R
\name{roboplotr_get_colors}
\alias{roboplotr_get_colors}
\title{Gives a vector of colors of the desired length.}
\usage{
roboplotr_get_colors(
  n_unique,
  robocolors = getOption("roboplot.colors.traces"),
  accessibility_params
)
}
\arguments{
\item{n_unique}{How many colors are needed.}

\item{robocolors}{Optional. A vector or colors.}

\item{accessibility_params}{Placeholder for accessibility features.}
}
\value{
plotly object

Vector of colors.
}
\description{
Outputs a vector of colors.
}
