% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_patterns.R
\name{roboplotr_get_pattern_showlegend}
\alias{roboplotr_get_pattern_showlegend}
\title{Determine if legend item for a given pattern is shown in \code{\link[=roboplot]{roboplot()}}}
\usage{
roboplotr_get_pattern_showlegend(d, pattern, pattern_showlegend)
}
\description{
Determine if legend item for a given pattern is shown in \code{\link[=roboplot]{roboplot()}}
}
