% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_patterns.R
\name{roboplotr_heatmap}
\alias{roboplotr_heatmap}
\title{interal function for adding heatmap styling \link{robotable}}
\usage{
roboplotr_heatmap(d, dt, heatmap)
}
\description{
interal function for adding heatmap styling \link{robotable}
}
