% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomap.R
\name{roboplotr_map_rasterlayer}
\alias{roboplotr_map_rasterlayer}
\title{Raster layer for data contour maps}
\usage{
roboplotr_map_rasterlayer(map, d, data_contour = F, opacity, robomap_palette)
}
\description{
Raster layer for data contour maps
}
