% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_patterns.R
\name{roboplotr_set_pattern}
\alias{roboplotr_set_pattern}
\title{Internal function for handling named patterns for traces}
\usage{
roboplotr_set_pattern(d, pattern, pattern_type)
}
\description{
Internal function for handling named patterns for traces
}
