% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomap.R
\name{roboplotr_set_zoom}
\alias{roboplotr_set_zoom}
\title{Internal setter for robomap zoom options}
\usage{
roboplotr_set_zoom(zoom, where = "in `robomap(zoom)`")
}
\description{
Internal setter for robomap zoom options
}
