% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotable.R
\name{robotable}
\alias{robotable}
\title{Automated datatable tables.}
\usage{
robotable(
  d,
  title = NULL,
  subtitle = "",
  caption = NULL,
  rounding = 1,
  width = getOption("roboplot.width"),
  height = NULL,
  flag = "",
  unit = "",
  pagelength = 10,
  info_text = NULL,
  heatmap = NULL,
  na_value = ""
)
}
\arguments{
\item{d}{Data frame. Data to be created a table from.}

\item{title, subtitle}{Characters. Labels for plot elements. Optionally, use \code{\link[=set_title]{set_title()}} for the title if you want to omit the title from the displayed plot, but include it for any downloads through the modebar.}

\item{caption}{Function or character. Use \code{\link[=set_caption]{set_caption()}}.}

\item{rounding}{Double. Controls the rounding of numeric columns.}

\item{height, width}{Double. Height and width of the table. Default width is NULL for responsive plots, give a value for static table dimensions.}

\item{flag, unit}{Character vectors. Use "+" for flag if you want to flag a numeric column for positive values. Unit is prepended to values in a numeric column. Name the units with column names from data 'd' if you want to target specific columns.}

\item{pagelength}{Double. Controls how many rows are displayed on the table. If data 'd' contains more rows than this, \code{\link[=robotable]{robotable()}} automatically adds navigation.}

\item{info_text}{Character. Optional. If included, this text will be displayed with a popup when the info button in the table modebar is clicked.}

\item{heatmap}{Function. Use \code{\link[=set_heatmap]{set_heatmap()}}. Displays any numeric values as a heatmap.}

\item{na_value}{Character. How NA values are displayed.}
}
\value{
A list of classes "datatable" and "htmlwidget"
}
\description{
Wrapper for \link[DT:datatable]{DT::datatable} for shorthand declaration of many layout arguments.
Automates many formatting options.
}
\examples{
# You can use roboplotr::robotable() to create html tables
#
energiantuonti |> robotable()
#
# No matter if you have any number date, character or numeric columns, robotable()
# handles the formatting automatically.
#
d <- energiantuonti |>
  dplyr::filter(Alue \%in\% c("Ruotsi","Kanada")) |>
  tidyr::unite(Tiedot, Alue, Suunta, sep = ", ") |>
  dplyr::arrange(Tiedot, time) |>
  tidyr::pivot_wider(names_from = Tiedot) |>
  dplyr::mutate(dplyr::across(where(is.numeric), ~ tidyr::replace_na(.x, 0))) |>
  dplyr::arrange(time)
d |> robotable()
#
# You can use the pagelength parameter to control how many rows are shown at once,
# info_text to add any information you wish describe about the table, and
# heatmap parameter if you want to color-code any numeric values across the
# table. See [set_heatmap()] documentation for further examples.
#
d |>
 robotable(heatmap = set_heatmap(),
           pagelength = 50,
           info_text = "Testing the info button.")

}
