% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{set_artefacts}
\alias{set_artefacts}
\title{Set global parameters in \code{\link[=roboplot_set_options]{roboplot_set_options()}} for artefact creation
of \code{\link[=roboplot]{roboplot()}} plots.}
\usage{
set_artefacts(
  artefacts = getOption("roboplot.artefacts")$artefacts,
  title = NULL,
  filepath = getOption("roboplot.artefacts")$filepath,
  render = getOption("roboplot.artefacts")$render,
  self_contained = getOption("roboplot.artefacts")$self_contained,
  auto = getOption("roboplot.artefacts")$auto
)
}
\arguments{
\item{artefacts}{Character vector. Controls what artefacts are saved. One or
more of "html", "img_w", "img_n" or "img_s". A matching button must be
present in the modebar (control modebar buttons with
\code{\link[=roboplot_set_options]{roboplot_set_options()}}.}

\item{title}{Character. The filename of the artefact(s) created
(without file format). Will be formatted with underscores, and the title
of argument 'p' will be used if no title is provided.}

\item{filepath}{Character. The filepath to the created artefacts.}

\item{render}{Logical. Controls if the plot saved will be displayed in
viewer. Plot will be returned silently in either case.}

\item{self_contained}{Logical. Controls if the plot dependencies will be
saved in an adjacent directory "plot_dependencies" or contained within the
file, increasing size.}

\item{auto}{Logical. Whether \code{\link[=roboplot]{roboplot()}} will create artefacts automatically.}
}
\value{
A list.
}
\description{
Set global parameters in \code{\link[=roboplot_set_options]{roboplot_set_options()}} for artefact creation
of \code{\link[=roboplot]{roboplot()}} plots.
}
\examples{
# Used to set global defaults for widget or other artefact creation. Any of
# these can be overridden by roboplotr::roboplot(). Only supposed to be
# called inside roboplotr::roboplot_set_options(). Use 'filepath' to control
# which directory the artefacts are created to, 'render' to control if the
# roboplot() plot will be rendered on artefact creation, 'self_contained' to
# control if html plot dependencies are placed in an adjacent directory or
# contained within the html file, and 'artefacts' (one of "html",
# "img_w","img_s", or "img_n") to control what artefacts are created.

# roboplot_create_widget() shows how the parameters are used. Further
# controls for other than "html" artefacts are are under
# set_imgdl_layout().
}
