% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{set_border}
\alias{set_border}
\title{Set global parameters in \code{\link[=set_roboplot_options]{set_roboplot_options()}} for plot borders
of \code{\link[=roboplot]{roboplot()}} plots.}
\usage{
set_border(
  xcolor = getOption("roboplot.border")$xcolor,
  ycolor = getOption("roboplot.border")$ycolor,
  xmirror = getOption("roboplot.border")$xmirror,
  ymirror = getOption("roboplot.border")$ymirror,
  xwidth = getOption("roboplot.border")$xwidth,
  ywidth = getOption("roboplot.border")$ywidth
)
}
\arguments{
\item{xcolor, ycolor}{Characters. Plot border colors. Must be a hexadecimal color strings or a valid css color strings.}

\item{xmirror, ymirror}{Logicals. Will the given axis be mirrored to left to right or bottom to top.}

\item{xwidth, ywidth}{Numerics. Plot border widths.}
}
\value{
A list.
}
\description{
Set global parameters in \code{\link[=set_roboplot_options]{set_roboplot_options()}} for plot borders
of \code{\link[=roboplot]{roboplot()}} plots.
}
\examples{
# You can control x- and y-axis border mirroring, width and color by using
# set_border() inside set_roboplot_options().

set_roboplot_options(
  border = set_border(ycolor = "green", xmirror = FALSE, xwidth = 5))

energiantuonti |> dplyr::filter(Alue == "Venäjä") |> roboplot(Suunta)

# Reset to defaults
set_roboplot_options(reset = TRUE)

}
