% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_menu.R
\name{set_external_menu}
\alias{set_external_menu}
\title{Set up an external menu for filtering categorical y-axis values}
\usage{
set_external_menu(
  p,
  btn = list(box = set_infobox(), font = set_font(color =
    getOption("roboplot.infobox")$color), labs = list(all = "Valitse kaikki", none =
    "Poista valinnat")),
  items = list(font = set_font(), checkbox = list(color =
    getOption("roboplot.colors.traces")[1], size = 1), width = list(min = 80, max = 180))
)
}
\arguments{
\item{p}{A plotly object created with \code{\link[=roboplot]{roboplot()}}.}

\item{btn}{A list defining the appearance and labels of the toggle button.}

\item{items}{A list defining the appearance and behavior of the checkbox items.}
}
\value{
A plotly object with an external menu for filtering categorical y-axis values.
}
\description{
Set up an external menu for filtering categorical y-axis values
}
\examples{
 # Basic usage with default settings
energiantuonti |>
  dplyr::filter(time == max(time)) |>
  roboplot(
    Suunta,
    plot_type = "bar",
    plot_mode = "horizontal",
    plot_axes = set_axes(x = "value", y = "Alue")
  ) |>
  set_external_menu()
}
