% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{set_grid}
\alias{set_grid}
\title{Plot grid for \code{\link[=roboplot]{roboplot()}}}
\usage{
set_grid(
  xcolor = getOption("roboplot.grid")$xcolor,
  ycolor = getOption("roboplot.grid")$ycolor,
  xwidth = getOption("roboplot.grid")$xwidth,
  ywidth = getOption("roboplot.grid")$ywidth,
  xdash = getOption("roboplot.grid")$xdash,
  ydash = getOption("roboplot.grid")$ydash
)
}
\arguments{
\item{xcolor, ycolor}{Characters. Plot gridline colors. Must be a hexadecimal color strings or a valid css color strings.}

\item{xwidth, ywidth}{Numerics. Plot gridline widths.}

\item{xdash, ydash}{Not currently supported by \link{plot_ly}. Characters. Plot gridline linetypes. Must contain one of "solid", "dash", "dot", "longdash", "dashdot", and "longdashdot" in any order.}
}
\value{
A list.
}
\description{
Set global parameters in \code{\link[=set_roboplot_options]{set_roboplot_options()}} for plot grid
of \code{\link[=roboplot]{roboplot()}} plots.
}
\examples{
# You can control x- and y-grid color, width and linetype by using
# set_grid() inside set_roboplot_options().

set_roboplot_options(
  grid = set_grid(ycolor = "green", ydash = "longdash", xwidth = 3))

energiantuonti |> dplyr::filter(Alue == "Venäjä") |> roboplot(Suunta)

# Reset to defaults
set_roboplot_options(reset = TRUE)

}
