% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modebar.R
\name{set_infobox}
\alias{set_infobox}
\title{Info box configuration for \code{\link[=roboplot]{roboplot()}} or \code{\link[=robotable]{robotable()}}}
\usage{
set_infobox(
  background = first(unlist(unique(getOption("roboplot.grid")[c("xcolor", "ycolor")]))),
  border = first(unlist(unique(getOption("roboplot.border")[c("xcolor", "ycolor")]))),
  border_width = 1
)
}
\arguments{
\item{background, border}{Characters. Colors used for the corresponding
element of the infopopup of \code{\link[=robotable]{robotable()}}s and \code{\link[=roboplot]{roboplot()}}s. Must be a
hexadecimal color or a valid css color.}

\item{border_width}{Integer. The border width of infomodal.}
}
\value{
A list of class roboplot.set_infobox.
}
\description{
Global parameters to add and customize info boxes in \link[=roboplot]{roboplots}
and \link[=robomap]{robomaps}.
}
