% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modebar.R
\name{set_modebar}
\alias{set_modebar}
\title{Modebar control for \code{\link[=roboplot]{roboplot()}} or \code{\link[=robotable]{robotable()}}}
\usage{
set_modebar(
  buttons = getOption("roboplot.modebar")$buttons,
  display = getOption("roboplot.modebar")$display,
  ...
)
}
\arguments{
\item{buttons}{Character vector. Buttons contained in modebar in the given order.
Must contain any of "home", "closest", "compare", "zoomin", "zoomout", "img_w",
"img_n", "img_s", "data_dl" and "robonomist" in any order.}

\item{display}{Character. One of "constant", "hover" or "none". Controls if the modebar
is visible always, only on hover, or never. Whatever the choice, static images will not
display modebar.}

\item{...}{Placeholder for other parameters.}
}
\value{
A list.
}
\description{
Set global parameters in \code{\link[=set_roboplot_options]{set_roboplot_options()}} for modebars.
}
