% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_robomap_legend}
\alias{set_robomap_legend}
\title{Legends for \code{\link[=robomap]{robomap()}}}
\usage{
set_robomap_legend(
  breaks = 5,
  labels = NULL,
  range = T,
  position = "bottomright",
  orientation = "vertical",
  title = NULL,
  opacity = 1
)
}
\arguments{
\item{breaks}{Numeric vector. A length 1 vector attempts to make that many legend entries. Length 2+ vector should be the breaks in values of param 'd' or \code{\link[=robomap]{robomap()}} how the legend should be split.}

\item{labels}{Character vector or numeric. If 'breaks' is of length 1, a numeric between 0 and 1, which is the rough proportion of legend entries \code{\link[=robomap]{robomap()}} will display (1 displays all). If 'breaks' is of length 2+, labels must be a character vector of length 1 + length of 'breaks'.}

\item{range}{Logical. Whether to use the maximum value of each legend item, or give a range. Ignored if 'labels' is provided.}

\item{position}{Character. One of 'bottomleft', 'bottomright', or 'none'.}

\item{orientation}{Character. Currently only accepts 'vertical'.}

\item{title}{Character. Currently unused.}

\item{opacity}{Numeric. The opacity of the legend, between 0 and 1.}
}
\value{
A list of class 'roboplotr.robomap.legend'.
}
\description{
Get a list for \code{\link[=robomap]{robomap()}} to parse the legend from.
}
