% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_roboplot_options}
\alias{set_roboplot_options}
\title{Override the default options for colors, fonts etc. for any plot created with \code{\link[=roboplot]{roboplot()}}}
\usage{
set_roboplot_options(
  accessible = NULL,
  artefacts = NULL,
  border = NULL,
  background_color = NULL,
  caption_defaults = NULL,
  dashtypes = NULL,
  font_main = NULL,
  font_title = NULL,
  font_caption = NULL,
  grid = NULL,
  height = NULL,
  imgdl_wide = NULL,
  imgdl_narrow = NULL,
  imgdl_small = NULL,
  linewidth = NULL,
  locale = NULL,
  logo_file = NULL,
  modebar = NULL,
  patterns = NULL,
  tick_colors = NULL,
  trace_colors = NULL,
  xaxis_ceiling = NULL,
  verbose = NULL,
  width = NULL,
  shinyapp = NULL,
  reset = F
)
}
\arguments{
\item{accessible}{Logical. Forces trace colors accessible against plot background.}

\item{artefacts}{Function. Control html and other file creation. Use \code{\link[=set_artefacts]{set_artefacts()}}.}

\item{border}{Function. Use \code{\link[=set_border]{set_border()}}.}

\item{background_color}{Character. Plot background color. Must be a hexadecimal color or a valid css color.}

\item{caption_defaults}{List. Used to parse caption. Values must be named "prefix", "lineend" and "updated". "prefix" is character, and added to caption text with ": ". "lineend" is character added to caption line ends. "updated" is logical that determines whether caption tries to guess latest update date from plot data.}

\item{dashtypes}{Character vector. Line trace linetypes in order of usage. Must contain all of "solid", "dash", "dot", "longdash", "dashdot", and "longdashdot" in any order.}

\item{font_main, font_title, font_caption}{Functions. Use \code{\link[=set_font]{set_font()}}.}

\item{grid}{Function. Use \code{\link[=set_grid]{set_grid()}}.}

\item{height, width}{Numerics. Height and width of roboplotr plots in pixels.}

\item{imgdl_wide, imgdl_narrow, imgdl_small}{Functions. Use \link{set_imgdl_layout}. Controls the dimensions and fonts of image files downloaded through modebar buttons.}

\item{linewidth}{Numeric. The default roboplotr line trace width.}

\item{locale}{Function. Defines locale parameters as \code{\link[=roboplot]{roboplot()}} needs them. Use \code{\link[=set_artefacts]{set_artefacts()}}.}

\item{logo_file}{Character. The filepath to the logo used in every plot.}

\item{modebar}{Character vector. Buttons contained in modebar in the given order. Must contain any of "home", "closest", "compare", "zoomin", "zoomout", "img_w", "img_n", "img_s", "data_dl" and "robonomist" in any order.}

\item{patterns}{Character vector. Line trace linetypes in order of usage. Must contain all of "", "/", "\\", "x", "-", "|", "+" and "." in any order.}

\item{tick_colors}{List. Plot tick element colors. Values need to be hexadecimal colors or valid css colors, named "x" and "y".}

\item{trace_colors}{Character vector. Trace colors in order of usage. Needs to be a hexadecimal color or a valid css color. You should provide enough colors for most use cases, while roboplotr adds colors as needed.}

\item{xaxis_ceiling}{Character. Default rounding for yaxis limit. One of "default", "days", "months", "weeks", "quarters", "years" or "guess".}

\item{verbose}{Character. Will roboplot display all messages, alerts and warnings, or warnings only? Must be one of "All", "Alert", or "Warning".}

\item{shinyapp}{Logical or list. Makes fonts, css and javascript available for shiny apps. If given as list, use a named list with the single character string named "container" describing the css selector for the element in a shiny app where most roboplots will be contained in. Used for relayouts if the plot is rendered while the container is not displayed.}

\item{reset}{Logical. Ignores other options, resets options to defaults.}
}
\description{
Override the default options for colors, fonts etc. for any plot created with \code{\link[=roboplot]{roboplot()}}
}
\examples{
# Control global options for roboplotr::roboplot(). Some of these you can set
# also in roboplotr::roboplot(), some are available only globally.


# Basic plot frame colors for ticks, grid and border must be defined by axis
# with lists of color hex codes or valid css colors. Same with plot backround.
# Height can be controlled by-plot or globally. You can also preconstruct
# some defaults for captions. Border control is further documented under
# set_border() function documentation.

d <- energiantuonti |>
  dplyr::filter(Alue \%in\% c("Kanada","Norja"), Suunta == "Tuonti")

set_roboplot_options(
  border = set_border(xcolor = "#eed5d2", ycolor = "#8b7d7b"),
  caption = list(prefix = "Lähde: ", lineend = ".", updated = FALSE),
  grid = set_grid(xcolor = "#9aff9a", ycolor = "cornsilk"),
  tick_colors = list(x = "darkgray", y = "dimgrey"),
  background_color = "ghostwhite",
  height = 700,
)

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")

p

# When testing different options, you can use 'reset' to reset default values.

set_roboplot_options(reset = TRUE)

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")

p

# You can set the displayed lower right logo by 'logo_file'. Logo file needs
# the filepath to the logo used. The logo will be automatically scaled for
# roboplotr::roboplot() usage. If the logo is not Robonomist logo, one will
# be automatically added to modebar (or you can add it manually as
# "robonomist").  Use 'verbose' to control if roboplotr writes all messages,
# alerts and warnings only, or warnings only.

# Of modebar options, "closest" and "compare" control on hover comparison
# points, and "zoomin" and "zoomout" are simple zoom buttons. You will
# probably want "home" along with the previous options for resetting the
# zoom. For downloading the plot data in .csv file format use "data_dl".
# Image sizes for downloaded image files are differentiated with
# "img_w"(ide), "img_n"(arrow) or "img_s"(mall).

set_roboplot_options(
  logo_file = system.file("images", "robonomist.png", package = "roboplotr"),
  modebar = c("home","closest","compare","zoomin","zoomout",
              "img_w","img_n","img_s","data_dl"),
  verbose = "All")

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")

p

# Images downloaded through plotly modebar require separate layout
# specifications which roboplotr::roboplot automatically takes care of. Image
# sizes and font sizes might require extra specifications. Set these globally
# with 'img_wide', 'img_narrow', and / or 'img_small' using
# roboplotr::set_imgdl_layout(), documented in detail in that
# function.

set_roboplot_options(imgdl_wide = set_imgdl_layout(width = 1600))


# Captions are partly controlled by 'caption defaults', while you must
# provide the basic text by-plot in roboplotr::roboplot().
# roboplotr::set_caption() is used on by-plot basis for more
# control, but you can provide global settings for some features. Provide a
# named list with all of "prefix", "lineend" and "updated", and captions will
# be changed accordingly. Use 'updated' = TRUE for roboplotr::roboplot() to
# try and extract the latest update from the data used.

set_roboplot_options(
  caption_defaults = list(prefix = "Lähteenä", lineend = "", updated = TRUE)
)

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus")

p

# Trace appearance is controlled globally by 'dashtypes', 'linewidth',
# 'patterns' and 'trace_colors'. Trace color you can determine by plot, the
# other options are available only globally. Linewidth simply controls the
# line width, but the other options are used specifically in the given order
# for roboplotr::roboplot() variables as factors. With only a few trace
# colors given roboplotr::roboplot() might struggle with a lot of traces, so
# it alerts the user for this. When provided, 'yaxis_ceiling' will leave
# either a predefined gap between x-axis end and the plot edge, or try to
# guess an appropriate gap. This can also be controlled by-plot and will not
# work with bar plots.

set_roboplot_options(
  dashtypes = c("longdash", "dashdot", "longdashdot", "solid", "dash", "dot"),
  linewidth = 4,
  patterns = c("x","-","|","+",".","","/","\\\\"),
  trace_colors = c("#027f93", "#153a42", "#f78b04"),
  xaxis_ceiling = "guess"
)

d <- energiantuonti |> dplyr::filter(Alue \%in\% c("Kanada","Norja"))

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus", pattern = Suunta)

p

p <- d |> roboplot(Alue, "Energian tuonti", "Milj €", "Tilastokeskus",
                   plot_type = "bar",
                   pattern = Suunta)

p

set_roboplot_options(reset = TRUE)


# When used inside shiny apps (assumed to use bs4Dash::dashboardPage()), run
# roboplotr::set_roboplot_options() in app ui header with 'shinyapp' = TRUE.

if(interactive()) {

ui <- bs4Dash::dashboardPage(
  bs4Dash::dashboardHeader(title = "Basic dashboard",
                           set_roboplot_options(shinyapp = TRUE)
  ),
  bs4Dash::dashboardSidebar(),
  bs4Dash::dashboardBody(
    # Boxes need to be put in a row (or column)
    shiny::fluidRow(
      bs4Dash::box(plotly::plotlyOutput("plot1", height = 550)),
      bs4Dash::box(plotly::plotlyOutput("plot2", height = 500))
    )
  )
)

server <- function(input, output) {

output$plot1 <- plotly::renderPlotly({
  roboplot(
    dplyr::filter(energiantuonti, Suunta == "Tuonti", Alue == "Kanada"),
    Alue, "Energian tuonti Kanadasta",
    "Milj. €",
    "Tilastokeskus")
})
output$plot2 <- plotly::renderPlotly({
  roboplot(
    dplyr::filter(energiantuonti, Suunta == "Vienti", Alue == "Kanada"),
    Alue, "Energian vienti Kanadaan",
    "Milj. €",
    "Tilastokeskus",
    height = 500)
})
}

shiny::shinyApp(ui, server)

}

# As roboplotr::roboplot() uses javascript to relayout the plot elements to
# correct sizes and placements on render, there might be cases where the user
# navigates between tabs too fast. This results in the plots rendering
# incorrectly as the element heights are zero at render time. Giving the
# parameter 'shinyapp' as a list with a character string named
# 'container' describing the css selector of the container that controls plot
# visibility allows roboplotr::roboplot() to detect if the container was
# displayed on plot rendering, redoing the incorrect layouts.

if(interactive()) {

  ui <- bs4Dash::dashboardPage(
    bs4Dash::dashboardHeader(
      title = "Basic dashboard",
      set_roboplot_options(
        shinyapp = list(container = ".tab-pane.container-fluid"))
    ),
    bs4Dash::dashboardSidebar(
      bs4Dash::sidebarMenu(
        htmltools::tagList(
          bs4Dash::menuItem(text = "Tab 1", tabName = "tab_1"),
          bs4Dash::menuItem(text = "Tab 2", tabName = "tab_2")
        )
      )
    ),
    bs4Dash::dashboardBody(
      # Boxes need to be put in a row (or column)
      uiOutput("plots")
    )
  )

  server <- function(input, output) {

    output$plots <- renderUI({
      bs4Dash::tabItems(
        bs4Dash::tabItem(tabName = "tab_1", shiny::fluidRow(
          bs4Dash::box(plotly::plotlyOutput("plot1", height = 550))
        )),
        bs4Dash::tabItem(tabName = "tab_2", shiny::fluidRow(
          bs4Dash::box(plotly::plotlyOutput("plot2", height = 500))
        ))
      )
    })

    output$plot1 <- plotly::renderPlotly({
      roboplot(
        dplyr::filter(energiantuonti, Suunta == "Tuonti", Alue == "Kanada"),
        Alue, "Energian tuonti Kanadasta",
        "Milj. €",
        "Tilastokeskus")
    })
    output$plot2 <- plotly::renderPlotly({
      roboplot(
        dplyr::filter(energiantuonti, Suunta == "Vienti", Alue == "Kanada"),
        Alue, "Energian vienti Kanadaan",
        "Milj. €",
        "Tilastokeskus",
        height = 500)
    })


  }

  shiny::shinyApp(ui, server)

}

}
