% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotable.R
\name{set_robotable_labels}
\alias{set_robotable_labels}
\title{Table label configuration}
\usage{
set_robotable_labels(
  search = getOption("roboplot.locale")$robotable_labels$search,
  info = getOption("roboplot.locale")$robotable_labels$info,
  lengthMenu = getOption("roboplot.locale")$robotable_labels$lengthMenu,
  emptyTable = getOption("roboplot.locale")$robotable_labels$emptyTable,
  first = getOption("roboplot.locale")$robotable_labels$first,
  last = getOption("roboplot.locale")$robotable_labels$last,
  `next` = ">>",
  previous = "<<"
)
}
\arguments{
\item{search}{Character. Label for search box.}

\item{info, lengthMenu, first, last, next, previous}{Character. Table navigation items.}

\item{emptyTable}{Character. The text shown when the table is empty.}
}
\description{
Parameters to add and customize labeling in \link[=robotable]{robotables}.
}
