% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotable.R
\name{set_table_options}
\alias{set_table_options}
\title{Set \code{robotable()} global styling defaults.}
\usage{
set_table_options(
  stripecolor = getOption("roboplot.table_defaults")$stripecolor,
  stripeopacity = getOption("roboplot.table_defaults")$stripeopacity,
  ...
)
}
\arguments{
\item{stripecolor}{Character. A valid css color name or 6-digit hex color.}

\item{stripeopacity}{Numeric. A number between 0 and 1.}

\item{...}{Placeholder for future table_options}
}
\description{
Control stripe appearance only at the moment.
}
