% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_title}
\alias{set_title}
\title{Titles for \code{\link[=roboplot]{roboplot()}}}
\usage{
set_title(title = NULL, include = T)
}
\arguments{
\item{title}{Character. Optional. Will try to determine a title based on
attributes of  argument 'd' of \code{\link[=roboplot]{roboplot()}}. Defaults to "PLACEHOLDER".}

\item{include}{Logical. Determines whether the given title will be used in
the plot. Will always inlcude it for downloaded png images.}
}
\value{
A list
}
\description{
Use in \code{\link[=roboplot]{roboplot()}} parameter 'title' to get a list used for setting the title.
}
\examples{
# Used to set titles for plots created with roboplotr::roboplot. You can
# simply give a charater string for plot titles.


d <- energiantuonti |> dplyr::filter(Alue == "Kanada",Suunta == "Tuonti")

d |> roboplot(Alue,"Energian tuonti Kanadasta")

# However, if you want to render the plot without the title included, you
# can use roboplotr::set_title(include = F) to omit it from the plot. This
# is for cases where you will insert the plot into environment where the
# title is outside the plot element. When downloading the plot, you would
# still want to have the title included, and roboplotr::set_plot() takes
# care of this. If you include a subtitle, it will still be in the plot.


d |>
roboplot(Alue,
title = set_title("Energian tuonti Kanadasta", include = FALSE),
subtitle = "Milj. €")


}
