% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeslider_zeroline.R
\name{set_zeroline}
\alias{set_zeroline}
\title{Zeroline configuration}
\usage{
set_zeroline(
  position = NULL,
  color = getOption("roboplot.zeroline")$color,
  width = getOption("roboplot.zeroline")$width
)
}
\arguments{
\item{position}{Numeric or logical. The position of the zeroline. If TRUE, the line is drawn at 0.}

\item{color}{Character. Must be a hexadecimal color or a valid css color.}

\item{width}{Numeric.}
}
\value{
A list of class roboplotr.set_zeroline
}
\description{
Parameters to customize zeroline in \link[=roboplot]{roboplots}.
}
\examples{
set_roboplot_options(zeroline = set_zeroline(color = "aliceblue", width = 4))
energiantuonti |> dplyr::filter(Alue == "Venäjä") |> roboplot(Suunta, zeroline = 200)
}
