% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{set_zoom}
\alias{set_zoom}
\title{Zooming configuration}
\usage{
set_zoom(zoom = getOption("roboplot.zoom"), ...)
}
\arguments{
\item{zoom}{Characters. Controls how \link[=roboplot]{roboplots} are zoomed in.
Valid strings are "none", "scroll" (scrollwheel and two-finger drag) and "drag"
(mouse or finger drag). Currently not used in \link[=robomap]{robomaps}. Default is "scroll".}

\item{...}{Placeholder for future arguments.}
}
\description{
Customize zooming behavior of \link[=roboplot]{roboplots}.
}
\examples{


# Zoom by dragging with mouse.
energiantuonti |>
  dplyr::filter(Alue == "Venäjä") |>
  roboplot(Suunta, zoom = "drag")

# Better for mobile to scroll with two fingers, or mobile on desktop.
energiantuonti |>
  dplyr::filter(Alue == "Venäjä") |>
  roboplot(Suunta, zoom = "scroll")

# Note that modebar zooming is not affected by this setting.
energiantuonti |>
  dplyr::filter(Alue == "Venäjä") |>
  roboplot(Suunta, zoom = "none",
           modebar = set_modebar(buttons = c("zoomin","zoomout","pan")))

}
