% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaesto_postinumeroittain}
\alias{vaesto_postinumeroittain}
\title{Energian tuonti ja vienti alkuper채maittain}
\format{
\subsection{\code{vaesto_postinumeroittain}}{

Kehys jossa 886 rivi채 and 5 saraketta:
\describe{
\item{Postinumeroalue}{Postinumeroalueen nimi}
\item{Alue}{Postinumeroalueen kunta}
\item{geom}{Monikulmiokohde}
\item{time}{Vuosi}
\item{value}{V채kiluku}
...
}
}
}
\source{
\url{https://pxdata.stat.fi/PxWeb/pxweb/fi/Postinumeroalueittainen_avoin_tieto/Postinumeroalueittainen_avoin_tieto__uusin/paavo_pxt_12f7.px/}
}
\usage{
vaesto_postinumeroittain
}
\description{
Otos postinumeroittaisesta v채kiluvusta Tilastokeskuksen Paavo-tilastotaulusta.
}
\keyword{datasets}
