% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energiantuonti}
\alias{energiantuonti}
\title{Energy Import and Export by Country of Origin}
\format{
\subsection{\code{energiantuonti}}{

Data frame with 723 rows and 4 columns:
\describe{
\item{Alue}{The name of the country or country group}
\item{Suunta}{Import or export}
\item{time}{Quarter}
\item{value}{Value (million euros)}
...
}
}
}
\source{
\url{https://pxdata.stat.fi/PxWeb/pxweb/fi/StatFin/StatFin__ehk/statfin_ehk_pxt_13j9.px/}
}
\usage{
energiantuonti
}
\description{
A sample of energy imports and exports by country of origin from Statistics Finland's archive table.
}
\keyword{datasets}
