% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomap.R
\name{robomap}
\alias{robomap}
\title{Automated leaflet maps.}
\usage{
robomap(
  d,
  area,
  title = NULL,
  caption = NULL,
  hovertemplate = list(flag = "", unit = ""),
  map_opacity = 0.7,
  map_colors = getOption("roboplot.colors.traces")
)
}
\arguments{
\item{d}{Data frame. Data to be created a table from.}

\item{area}{Symbol, string, or function resulting in symbol or string. Variable from argument 'd' to use to identify the areas described by the map. This must be of class sfc_MULTIPOLYGON.}

\item{title, subtitle}{Characters. Labels for plot elements. Optionally, use \code{\link[=set_title]{set_title()}} for the title if you want to omit the title from the displayed plot, but include it for any downloads through the modebar.}

\item{map_opacity}{Double. Value from 0 to 1, defining how opaque the underlying map tiles are. 0 removes the map layer.}

\item{map_colors}{Character vector. Colors used for map tiles, from highest to lowest values. \code{\link[=robomap]{robomap()}} will extrapolate as many colors as necessary. Must be a hexadecimal color strings or a valid css color strings.}
}
\value{
A list of classes "leaflet" and "htmlwidget"
}
\description{
Wrapper for \link[leaflet:leaflet]{leaflet::leaflet} for shorthand declaration of many map layout arguments.
Automates many formatting options.
}
\examples{
# You can use roboplotr::robomap() to create html maps. Note that very large
# number of geoms makes for slow rendering maps.

vaesto_postinumeroittain |> robomap(Postinumeroalue)

# Currently robomap() only supports very little customization.
vaesto_postinumeroittain |>
  dplyr::filter(Alue == "Espoo") |>
  robomap(Postinumeroalue, title = "Väestö Espoossa", caption = "Tilastokeskus")

# robomap() adjusts legend item rounding based on the largest value in the "value"
# column of data 'd'. Control trace colors currently with global colors only.

set_roboplot_options(trace_colors = c("lightgreen","darkred"), trace_border = list(width = 2, color = "black"))

vaesto_postinumeroittain |>
  dplyr::slice_min(value, n = 30) |>
  robomap(Postinumeroalue, title = "Vähäväkilukuiset postinumerot", caption = "Tilastokeskus")

}
