% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{roboplot}
\alias{roboplot}
\title{Responsive plotly charts for embedding in shiny applications or websites}
\usage{
roboplot(
  d,
  color = NULL,
  title = NULL,
  subtitle = "",
  caption = NULL,
  legend_orientation = NULL,
  legend_position = NULL,
  margin = NA,
  trace_color = NULL,
  highlight = NULL,
  zeroline = F,
  rangeslider = FALSE,
  pattern = NULL,
  line_width = getOption("roboplot.linewidth"),
  hovertext = NULL,
  axis_limits = list(x = c(NA, NA), y = c(NA, NA)),
  plot_type = "scatter",
  plot_mode = "line+dodge",
  plot_yaxis,
  height = getOption("roboplot.height"),
  facet_split = NULL,
  pie_rotation = F,
  legend_maxwidth = NULL,
  xaxis_ceiling = c("default", "days", "months", "weeks", "quarters", "years", "guess"),
  secondary_yaxis = NULL
)
}
\arguments{
\item{d}{Data frame. Data to be plotted.}

\item{color}{Expression. Variable from argument 'd' to use for trace color. If left NULL, the argument 'subtitle' will be used as a placeholder for determining color and hoverlabels.}

\item{title, subtitle}{Characters. Labels for plot elements.}

\item{caption}{Function or character. Use roboplot_set_caption().}

\item{legend_orientation, legend_position}{Characters. Currently only "bottom" or NA for no legend.}

\item{trace_color}{Character vector, named if length > 1. Trace color for all trace. Determines the trace type for either the whole plot, or for all variables defined by color as name-value pairs.}

\item{highlight}{Double or list. Determines if a given trace is included in legend and assigned a color.
If double, traces with max(value) < highlight will be give trace color matching the grid color, and removed from the legend.
If function, it must return a logical and include named items "value" and ".fun", where .fun checks if given value will get a color or legend item.}

\item{zeroline}{Logical or double. Determines zeroline inclusion, TRUE for zeroline, or double for exact placement.}

\item{rangeslider}{Logical or character in \%Y-\%m-\%d format. Determines rangeslider inclusion. TRUE includes the rangeslider, a character string includes the rangeslider with the given date as a start date.}

\item{pattern}{Expression. Variable from argument 'd' to use for linetype or bar pattern. Not supported for bar charts.}

\item{line_width}{Double vector, named if length > 1. Line width for all line traces. Determines the line width for either the whole plot, or for all variables defined by color as name-value pairs.}

\item{hovertext}{Function. Use roboplot_get_hovertemplate_specs().}

\item{axis_limits}{List. Determines the limits of the axes (list(x = c(NA,NA), y = c(NA,NA)))".}

\item{plot_type}{Character vector, named if length > 1. Determines the trace type for either the whole plot, or for all variables defined by color as name-value pairs.}

\item{plot_mode}{Character. Determines the barmode for a barplot. Disregarded for scatterplots.}

\item{plot_yaxis}{Expression. Variable from argument 'd' to use as y-axis for a horizontal barplot. Disregarded for other plots.}

\item{height}{Double. Height of the plot.}

\item{facet_split}{Currently unused. Variable from argument 'd' to use for facet splits.}

\item{pie_rotation}{Logical. Controls if 0°-mark of a pie is centered on middle of the first item of the color variable as factor.}

\item{legend_maxwidth}{Double. Legend items (and y-axis values for horizontal barplots) longer than this will be collapsed with an ellipsis (Double).}

\item{xaxis_ceiling}{Character. One of "default", "days", "months", "weeks", "quarters", "years", or "guess"). How to round the upper bound of plot x-axis for other than bar plots if no axis limits are given.}

\item{secondary_yaxis}{Expression. Variable from argument 'd' resulting in a maximum of two factor levels, determining which observations if any use a secondary y-axis.
Parameter 'zeroline' will be ignored. Cannot currently differentiate between the axes in legend, and right margin will not scale properly on zoom and possibly on png generation.}
}
\value{
A list object of classes "plotly" and "html"
}
\description{
Wrapper for plotly::plot_ly for shorthand declaration of many layout and trace arguments.
Included relayout.js ensures proper scaling or elements when used in shiny apps, iframes, static image downloads and so on.
}
\examples{
# Basic usage is for plotting lines
d <- energiantuonti |>
   dplyr::filter(Alue \%in\% c("Kanada","Norja","Yhdistynyt kuningaskunta"))
d1 <- d |> dplyr::filter(Suunta == "Tuonti")
p <- d1 |> roboplot(Alue,"Energian tuonti","Milj. €","Euroopan keskuspankki")

# Pattern can be used for lines and ordering your variables to factors
# will affect the order in which the traces are added and presented.
d2 <- d |> dplyr::mutate(Alue = forcats::fct_reorder(Alue, value))
p <- d2 |>
   roboplot(Alue,"Energian tuonti ja vienti","Milj. €","Euroopan keskuspankki",
   pattern = Suunta)

# Bar charts use pattern too
p <- d2 |>
   roboplot(Alue,"Energian tuonti ja vienti","Milj. €","Euroopan keskuspankki",
   pattern = Suunta,
   plot_type = "bar")

# With single 'time' observation x-axis tickmarks lose tick labels
d3 <- d2 |> dplyr::filter(time == max(time))
p <- d3 |>
   roboplot(Alue,"Energian tuonti ja vienti","Milj. €","Euroopan keskuspankki",
   pattern = Suunta,
   plot_type = "bar")

# Bar chart can be horizontal
# but then is better off with only a single 'time' observation
p <- d3 |>
   roboplot(Alue,"Energian tuonti ja vienti","Milj. €","Euroopan keskuspankki",
   pattern = Suunta,
   plot_type = "bar",
   plot_mode = "horizontal",
   plot_yaxis = "Alue")

# Pie charts are possible too, but pattern is currently ignored by plotly.
p <- d3 |>
   roboplot(Alue,"Energian tuonti ja vienti yhteensä","Milj. €","Euroopan keskuspankki",
   pattern = Suunta,
   plot_type = "pie")

# Pie chart can be rotated to center the first factor level of 'color'.
p <- d3 |>
   roboplot(Alue,"Energian tuonti ja vienti yhteensä","Milj. €","Euroopan keskuspankki",
   pattern = Suunta,
   plot_type = "pie",
   pie_rotation = T)
}
