% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{roboplot_create_widget}
\alias{roboplot_create_widget}
\title{Writes an html element for embedding, and optionally png files.}
\usage{
roboplot_create_widget(
  p,
  title,
  filepath,
  render = T,
  self_contained = F,
  png_artefacts
)
}
\arguments{
\item{p}{A plotly object.}

\item{title}{The filename of the html element (without file format). The function will clean the name up, or try to extract it from param p if missing.}

\item{filepath}{The path of the saved file.}

\item{render}{Logical. Is the plot rendered to viewer after saving the widget (default true). Returns the plot object nonetheless.}

\item{self_contained}{Logical. Will the html artefact have self-contained dependencies, increasing size. Default false.}

\item{png_artefacts}{Optional character vector of s(mall), n(arrow), and/or w(ide) corresponding to the expected .png sizes.}
}
\value{
The plotly object p.
}
\description{
Writes an html element for embedding, and optionally png files.
}
\examples{
\dontrun{
p |> roboplot_create_widget()
}
}
