% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{roboplot_set_caption}
\alias{roboplot_set_caption}
\title{Get a string for \code{\link[=roboplot]{roboplot()}} captions.}
\usage{
roboplot_set_caption(
  text = NULL,
  prefix = getOption("roboplot.caption")$prefix,
  updated = getOption("roboplot.caption")$updated,
  .data = NULL,
  line.end = getOption("roboplot.caption")$lineend,
  prepend = NULL,
  append = NULL
)
}
\arguments{
\item{text}{Character. The text for plot caption, probably data source name for argument 'd'.}

\item{prefix}{Character. Prefix inserted before the argument 'text', automatically separated by ": ".}

\item{updated}{Date or logical. If Date, will be inserted as the last update time in the caption. If TRUE, will try to get this information from argument '.data'.}

\item{.data}{Data frame. The data frame the caption attempts to read the time of the last update from.}

\item{line.end}{Character. Character inserted to end of line. Default from roboplot.options.}

\item{append, prepend}{Character vectors. Characters inserted after or before other caption text. Every item in vector will get a new line.}
}
\value{
A string.
}
\description{
Get a string for \code{\link[=roboplot]{roboplot()}} captions.
}
\examples{
# Used to define how captions are constructed inside roboplotr::roboplot()
# captions. 'Text' is the caption text, and if only this is needed, you can
# use simply write the string. roboplotr::roboplot() will provide the
# prefix and end-of-line character from global options that can be altered
# with roboplotr::roboplot_set_options().

d <- energiantuonti |> dplyr::filter(Alue == "Kanada",Suunta == "Tuonti")
p <- d |> roboplot(Alue,
                   title = "Energian tuonti Kanadasta",
                   subtitle = "Miljoonaa euroa",
                   caption = "Tilastokeskus"
)

p

# If needed, the individual elements can be entered as function parameters
# or simply provide everything as strings, overriding the defaults with
# roboplotr::roboplot_set_options().

p <- d |>
  roboplot(Alue,
           title = "Energy import",
           subtitle = "Million euros",
           caption = roboplot_set_caption(
             prepend = "From Canada",
             append = paste0("(Customs Finland, International trade ",
                            "statistics;\nRadiation and Nuclear Safety ",
                            "Authority; Gasum LLC)"),
             text = "Statistics Finland",
             prefix = "Source",
             line.end = ""
           )
  )

p

# If 'updated' is set to TRUE, need to provide the .data where
# roboplotr::roboplot_set_caption() will look for the update info, and if
# found, it appears under the caption text.
p <- d |> roboplot(Alue,
                   title = "Energian tuonti Kanadasta",
                   subtitle = "Miljoonaa euroa",
                   caption = roboplot_set_caption(
                     text = "Tilastokeskus",
                     updated = TRUE,
                     .data = d
                   )
)

p

# If you need make manual changes repeatedly, you are probably better off
# using roboplotr::roboplot_set_options() to change the defaults to something
# more sensible.
}
