% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{roboplot_set_font}
\alias{roboplot_set_font}
\title{Creates a list used to construct plot fonts}
\usage{
roboplot_set_font(
  path = "sans-serif",
  size = 12,
  color = "black",
  bold = T,
  type = NULL
)
}
\arguments{
\item{path}{Character. Either "serif", "sans-serif" or a path to .otf or .ttf file.}

\item{size}{Double. Determines font size for the font.}

\item{color}{Character. Must be a hexadecimal color or among strings provided by grDevices::colors.}

\item{bold}{Logical. Only used for title font. Determines if the title is bolded.}

\item{type}{Character. One of "main", "title" or "caption".}
}
\value{
A list
}
\description{
Creates a list used to construct plot fonts
}
