% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roboplot_set_options}
\alias{roboplot_set_options}
\title{Override the default options for colors, fonts etc. for any plot created with roboplot()}
\usage{
roboplot_set_options(roboplot_options, notify = T, shinyapp = F)
}
\arguments{
\item{roboplot_options}{List. The roboplot options to be set.}

\item{notify}{Logical. Controls the message for set options.}

\item{shinyapp}{Logical. Makes fonts, css and javascript available for shiny apps.}
}
\description{
Override the default options for colors, fonts etc. for any plot created with roboplot()
}
