% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{set_artefacts}
\alias{set_artefacts}
\title{Visualization export configuration}
\usage{
set_artefacts(
  artefacts = getOption("roboplot.artefacts")$artefacts,
  title = NULL,
  filepath = getOption("roboplot.artefacts")$filepath,
  render = getOption("roboplot.artefacts")$render,
  self_contained = getOption("roboplot.artefacts")$self_contained,
  zoom = getOption("roboplot.artefacts")$zoom,
  auto = getOption("roboplot.artefacts")$auto,
  width = getOption("roboplot.artefacts")$width,
  height = getOption("roboplot.artefacts")$height,
  delay = getOption("roboplot.artefacts")$delay,
  alt = getOption("roboplot.artefacts")$alt
)
}
\arguments{
\item{artefacts}{Character vector. Controls what artefacts are saved. One or
more of "html", "png", "jpg", "jpge", "webp", or "pdf".}

\item{title}{Character. The filename of the artefact(s) created
(without file format). Will be formatted with underscores, and the \code{title} text
of \code{p} will be used if no title is provided.}

\item{filepath}{Character. The filepath to the created artefacts.}

\item{render}{Logical. Controls if the artefact will be displayed in
viewer. Will be returned silently in either case.}

\item{self_contained}{Logical. Controls whether artefact's dependencies' are
saved in an adjacent directory or contained within the
file, increasing size.}

\item{zoom}{Numeric. Controls the zoom level of static plots if any are
defined with 'artefacts'. Default 1.}

\item{auto}{Logical. Whether \link[=roboplot]{roboplot} or \link[=robotable]{robotable}
will create artefacts automatically.}

\item{width, height}{Numeric. Sets the size of any static plots created. Any
artefacts created with \code{\link[=roboplot]{roboplot()}}'s \code{artefacts} parameter will use the given
dimensions, if any, for that plot.}

\item{delay}{Numeric. Delay in milliseconds before taking a screenshot. Used with static file creation. Default 0.2}

\item{alt}{Logical. Will a html artefact have the param \code{title} as its \code{alt}
attribute.}
}
\value{
A list of class roboplot.set_artefacts.
}
\description{
Parameters to configure exports from \link[=roboplot]{roboplots} or \link[=robotable]{robotables}.
}
\examples{
# Used to set global defaults for widget or other artefact creation. Any of
# these can be overridden by `roboplot()`. Use `filepath` to control which
# directory the artefacts are created to, `render` to control if the object will
# be rendered on artefact creation, `self_contained` to control if html plot
# dependencies are placed in an adjacent directory or contained within the html file,
# `artefacts` (one of "html", "png","jpg", "jpge", "jpge", "webp" or "pdf) to
# control what artefacts are created, and 'zoom' to set static artefact zoom level.

# create_widget() shows how the parameters are used.
}
