% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_caption}
\alias{set_caption}
\title{Get a string for \code{\link[=roboplot]{roboplot()}} captions.}
\usage{
set_caption(
  text = NULL,
  prefix = getOption("roboplot.caption")$prefix,
  updated = getOption("roboplot.caption")$updated,
  .data = NULL,
  line.end = getOption("roboplot.caption")$lineend,
  prepend = NULL,
  append = NULL
)
}
\arguments{
\item{text}{Character. The text for plot caption, probably data source name for argument 'd' of \code{\link[=roboplot]{roboplot()}}.}

\item{prefix}{Character. Prefix inserted before the argument 'text'. The prefix will automatically separated from text by ": ".}

\item{updated}{Date or logical. If Date, will be inserted as the last update time in the caption. If TRUE, will try to get this information from argument '.data'.}

\item{.data}{Data frame. The data frame the caption attempts to read the time of the last update from. Use the same data you are providing as argument 'd' of \code{\link[=roboplot]{roboplot()}}.}

\item{line.end}{Character. Character inserted to end of line. Default from roboplot.options.}

\item{append, prepend}{Character vectors. Characters inserted after or before other caption text. Every item in vector will get a new line.}
}
\value{
A string.
}
\description{
Get a string for \code{\link[=roboplot]{roboplot()}} captions.
}
\examples{
# Used to define how captions are constructed inside roboplotr::roboplot()
# Use 'text' as the caption text, and if only this is needed, you can
# use simply write the string. roboplotr::roboplot() will provide the
# prefix and end-of-line character from global options that can be altered
# with roboplotr::set_roboplot_options().


d <- energiantuonti |> dplyr::filter(Alue == "Kanada",Suunta == "Tuonti")

d |> roboplot(Alue,"Energian tuonti Kanadasta","Milj. €",
                   caption = "Tilastokeskus")


# Override the global options with function parameters.

d |>
  roboplot(Alue, "Energy import","Million euros",
           caption = set_caption(
             prepend = "(Canada)",
             append = paste0("(Customs Finland, International trade ",
                             "statistics;\nRadiation and Nuclear Safety ",
                             "Authority; Gasum LLC)"),
             text = "Statistics Finland",
             prefix = "Source: ",
             line.end = "")
  )


# If 'updated' is set to TRUE, you need to provide the .data where
# roboplotr::set_caption() will look for the update info, and if
# found, it appears under the caption text.
d |> roboplot(Alue,"Energian tuonti Kanadasta","Miljoonaa euroa",
              caption = set_caption(
                text = "Tilastokeskus",
                updated = TRUE,
                .data = d
              )
)

# If you need to make manual changes repeatedly, you are probably better off
# using roboplotr::set_roboplot_options() (documented therein) to change the
# defaults to something more sensible.

set_roboplot_options(
  caption_defaults = list(prefix = "Source: ", lineend = "", updated = FALSE)
  )

d |> roboplot(Alue,"Energy import from Canada","M€", "Statistic Finland")

# Revert to defaults:
set_roboplot_options(reset = TRUE)

}
