% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{set_errorbars}
\alias{set_errorbars}
\title{Error bar configuration}
\usage{
set_errorbars(...)
}
\arguments{
\item{...}{Parameters passed to \code{set_confidence_interval()}.}
}
\value{
A list of class roboplot.set_confidence_interval
}
\description{
Parameters to add and customize error bars in \link[=roboplot]{roboplots}.
}
