% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{set_errorbars}
\alias{set_errorbars}
\title{Set the error bars}
\usage{
set_errorbars(
  error_x = NULL,
  error_y = NULL,
  xcolor = getOption("roboplot.errorbars")$xcolor,
  ycolor = getOption("roboplot.errorbars")$xcolor
)
}
\arguments{
\item{error_x, error_y}{Symbol, string, or function resulting in symbol or
string. Columns from param 'd' or \code{\link[=roboplot]{roboplot()}} error bars.}

\item{xcolor, ycolor}{Characters. Error bar colors. Must be a hexadecimal color
strings or a valid css color strings.}
}
\description{
Set the error bars
}
\examples{
# You need a column for error bar values to add error bars. Then refer to that
# numeric column by set_errorbars() with the proper axis.
d <- energiantuonti |>
  dplyr::filter(Suunta == "Tuonti", Alue == "Venäjä", year(time) > 2016) |>
  dplyr::group_by(Alue) |>
  dplyr::mutate(sd = sd(value)) |>
  dplyr::ungroup()
d |> roboplot(Alue,
              plot_type = "bar",
              error_bars = set_errorbars(error_y = sd))
# You can change the color by providing the proper reference.
d |> roboplot(Alue,
              plot_type = "bar",
              error_bars = set_errorbars(error_y = sd, ycolor = "green"))

# Remember to refer to proper axis.
d |> roboplot(Alue,
              plot_type = "bar",
              plot_mode = "horizontal",
              plot_axes = set_axes(y = "time", yticktype = "date"),
              error_bars = set_errorbars(error_x = sd, xcolor = "pink"))


}
