% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_font}
\alias{set_font}
\title{Get a list for \code{\link[=set_roboplot_options]{set_roboplot_options()}} used as font specifications in \code{\link[=roboplot]{roboplot()}}}
\usage{
set_font(
  font = "Arial",
  fallback = NULL,
  size = 12,
  color = "black",
  bold_title = T,
  type = NULL
)
}
\arguments{
\item{font}{Character. One of "Arial", "Verdana", "Tahoma", "Trebuchet", "Times New Roman", "Georgia", "Garamond", "Courier New", "Brush Script MT" or a path to an .otf or .ttf file.}

\item{fallback}{Character. Only used when a file path is used with 'font'. Fallback font when the defined font is unavailable. Especially concerns image downloads.}

\item{size}{Double. Determines font size for the font.}

\item{color}{Character. Must be a hexadecimal color or a valid css color.}

\item{bold_title}{Logical. Only used for font of type 'title'. Determines if the title is bolded.}

\item{type}{Character. One of "main", "title" or "caption".}
}
\value{
A list
}
\description{
Get a list for \code{\link[=set_roboplot_options]{set_roboplot_options()}} used as font specifications in \code{\link[=roboplot]{roboplot()}}
}
\examples{
# Used to set fonts used by roboplotr::roboplot(). Only supposed to be called
# inside roboplotr::set_roboplot_options(). 'path' can be a file path or
# string "serif" or "sans-serif", if no specific font path will be provided.
# This will change fonts according to the list output of
# roboplotr::set_font(). You can designate font color, size, and
# path, and (for title font only) whether it is bolded. You must also provide
# one of "title", "main" or "caption" as 'type' and a web-safe fallback font
# if an actual filepath is
# provided for 'path'.


set_roboplot_options(
  caption = list(prefix = "Lähde: ", lineend = ".", updated = FALSE),
  font_title = set_font(size = 17, font = "Tahoma", bold = TRUE),
  font_main =  set_font(
    size = 11,
    font = system.file("www","fonts","bladerunner.TTF", package = "roboplotr"),
    fallback = "Verdana",
    color = "darkred",
    type = "main"),
  font_caption = set_font(color = "green")
)

d <- energiantuonti |> dplyr::filter(Alue == "Kanada",Suunta == "Tuonti")

d |> roboplot(Alue,"Energian tuonti Kanadasta","Milj. €","Tilastokeskus")

# Note that plotly's own downloadImage does not support external fonts.

if(interactive()) {

  d |>
  roboplot(Alue,"Energian tuonti Kanadasta","Milj. €","Tilastokeskus") |>
  roboplot_create_widget(filepath = tempdir(), artefacts = "img_w")

  utils::browseURL(paste0(tempdir(),"/energian_tuonti_kanadasta_levea.png"))
}

# revert to defaults:
set_roboplot_options(reset = TRUE)
}
