% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hovertemplate.R
\name{set_hovertext}
\alias{set_hovertext}
\title{Hovertext configuration}
\usage{
set_hovertext(
  frequency = NULL,
  rounding = getOption("roboplot.rounding"),
  unit = "",
  text_col = NULL,
  extra = NULL
)
}
\arguments{
\item{frequency}{Character. Determines the format of hovertemplate dates. One
of "Annual", "Quarterly", "Monthly", "Weekly", "Daily", or NULL. Defaults to
NULL, where \code{\link[=roboplot]{roboplot()}} infers the format from \code{d}.}

\item{rounding}{Double. Number of decimal places for hovertemplate values.}

\item{unit}{Character. Unit displayed.}

\item{text_col}{Symbol or character. Column used from \code{d} in a \link[=roboplot]{roboplot}
for labeling. If NULL, the column used for \code{color} (and, if given, \code{pattern})
is used for labels.}

\item{extra}{Character vector. Extra text displayed under any hovertext.}
}
\value{
A list of class roboplotr.set_hovertext
}
\description{
Parameters to control the formatting and content of hovertext in \link[=roboplot]{roboplots}.
}
\examples{
# Use to give hovertemplate specifications for `roboplot()` plots.
# Without specifying hovertext, `roboplot()` will construct it based
# on data from argument `d` of `roboplot()`, guessing frequency and using
# the argument `subtitle` as the unit of value. Control formatting with
# `frequency`, `rounding`, and `unit`.

d <- energiantuonti |>
  dplyr::filter(Alue \%in\% c("Kanada","Norja","Yhdistynyt kuningaskunta"),
                Suunta == "Tuonti")

d |> roboplot(
  Alue, "Energian tuonti", "Miljoonaa euroa", "Lähde: Tilastokeskus.",
  hovertext = set_hovertext(
    frequency = "Monthly", rounding = 2, unit = "Milj. €"
    )
)

# Parameter 'extra' will be added to last line(s) of the hovertemplate, which
# can be plain text or reference to `d` of the `roboplot()`, with d3 syntax.

d |> roboplot(
  Alue, "Energian tuonti", "Milj. €", "Lähde: Tilastokeskus.",
  hovertext = set_hovertext("Annual", extra = "(\%{x:|Q})")
)

}
