% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modebar.R
\name{set_infobox}
\alias{set_infobox}
\title{Infobox appearance control for \code{\link[=roboplot]{roboplot()}} or \code{\link[=robotable]{robotable()}}}
\usage{
set_infobox(
  background = first(unlist(unique(getOption("roboplot.grid")[c("xcolor", "ycolor")]))),
  border = first(unlist(unique(getOption("roboplot.border")[c("xcolor", "ycolor")]))),
  border_width = 1
)
}
\arguments{
\item{background, border}{Characters. Colors used for the corresponding
element of the infopopup of \code{\link[=robotable]{robotable()}}s and \code{\link[=roboplot]{roboplot()}}s. Must be a
hexadecimal color strings or a valid css color strings.}

\item{border_width}{Integer. The border width of infomodal.}
}
\value{
A list.
}
\description{
Set global parameters in \code{\link[=set_roboplot_options]{set_roboplot_options()}} for plot borders
of \code{\link[=roboplot]{roboplot()}} plots and \code{\link[=robotable]{robotable()}} tables.
}
