% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_locale}
\alias{set_locale}
\title{Locales for \code{\link[=roboplot]{roboplot()}}}
\usage{
set_locale(locale = "fi-FI")
}
\arguments{
\item{locale}{Character. Currently supports on only "en-GB", "en-US",
"sv-SE", or "fi-FI" (the default).}
}
\value{
A list
}
\description{
Used to set locale parameters for \code{\link[=roboplot]{roboplot()}} in \code{\link[=set_roboplot_options]{set_roboplot_options()}}
}
\examples{
# You might want to display dates or numbers with some another default format
# for decimal marks or thousand marks.

set_roboplot_options(
  locale = set_locale("en-GB"),
  caption_template = "Source: {caption}."
  )

d <- energiantuonti |>
  dplyr::filter(Alue == "Kanada", Suunta == "Tuonti") |>
  dplyr::mutate(value = value * 1000000, Alue = "Canada")

d |>
  roboplot(color = Alue,
           title = "Energy import",
           subtitle = "Mil. \u20AC",
           caption = "Statistics Finland")

# This works on dates, too, if you need to show an exact date. In this
# example you wouldn't, and it only matters on the hoverlabel here, but let's
# force roboplotr() to display dates at by-day accuracy for the sake of example
# by using data frequency of "Daily".

attr(d, "frequency") <- "Daily"

d |> roboplot(color = Alue,
              title = "Energy import",
              subtitle = "Mil. \u20AC",
              caption = "Statistics Finland")

# Revert to defaults:
set_roboplot_options(reset = TRUE)

}
