% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modebar.R
\name{set_modebar}
\alias{set_modebar}
\title{Modebar configuration}
\usage{
set_modebar(
  buttons = getOption("roboplot.modebar")$buttons,
  display = getOption("roboplot.modebar")$display,
  title = getOption("roboplot.modebar")$title,
  ...
)
}
\arguments{
\item{buttons}{Character vector. Buttons contained in modebar in the given order.
Must contain any of "home", "closest", "compare", "zoomin", "zoomout", "img_w",
"img_n", "img_s", "data_dl" and "robonomist" in any order.}

\item{display}{Character. One of "constant", "hover" or "none". Controls whether
modebar is visible always, only on hover, or never. Whatever the choice, static
exports will not display a modebar.}

\item{...}{Placeholder for other parameters.}

\item{title.}{This will be the title used for static files downloaded through
modebar. Control image download title extension specifications with \code{set_imgdl_layout()}.
\link[=roboplot]{roboplot} \code{title} will be used if no title is given here.}
}
\value{
A list.
}
\description{
Parameters to add and customize modebars in \link[=roboplot]{roboplots}.
}
