% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotable.R
\name{set_robotable_labels}
\alias{set_robotable_labels}
\title{Set labels for robotable items like search box and pagination.}
\usage{
set_robotable_labels(
  search = "Etsi:",
  info = "Näytetään rivit _START_-_END_ / _TOTAL_",
  lengthMenu = "Näytä _MENU_ riviä per sivu",
  emptyTable = "Tietoja ei saatavilla",
  first = "Ensimmäinen",
  last = "Viimeinen",
  `next` = ">>",
  previous = "<<"
)
}
\arguments{
\item{search}{Character. Label for search box.}

\item{info, lengthMenu, first, last, next, previous}{Character. Table navigation items.}

\item{emptyTable}{Character. The text shown when the table is empty.}
}
\description{
Set labels for robotable items like search box and pagination.
}
