% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeslider_zeroline.R
\name{set_zeroline}
\alias{set_zeroline}
\title{Set global defaults for \code{\link[=roboplot]{roboplot()}} zerolines
Only relevant when a zeroline is set with \code{\link[=roboplot]{roboplot()}}}
\usage{
set_zeroline(
  position,
  color = getOption("roboplot.zeroline")$color,
  width = getOption("roboplot.zeroline")$width
)
}
\arguments{
\item{position}{Numeric. Zeroline location.}

\item{color}{Character. Zeroline color. Must be a hexadecimal color strings or a valid css color strings.}

\item{width}{Numeric. Zeroline width.}
}
\description{
Set global defaults for \code{\link[=roboplot]{roboplot()}} zerolines
Only relevant when a zeroline is set with \code{\link[=roboplot]{roboplot()}}
}
\examples{
set_roboplot_options(zeroline = set_zeroline(color = "aliceblue", width = 4))
energiantuonti |> dplyr::filter(Alue == "Venäjä") |> roboplot(Suunta, zeroline = 200)
}
