% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaesto_postinumeroittain}
\alias{vaesto_postinumeroittain}
\title{Population by Postal Code Area}
\format{
\subsection{\code{vaesto_postinumeroittain}}{

Data frame with 886 rows and 6 columns:
\describe{
\item{geom}{Multipolygon object}
\item{Postinumero}{Postal Code}
\item{Postinumeroalue}{Name of the Postal Code Area}
\item{Alue}{Municipality of the Postal Code Area}
\item{time}{Year}
\item{value}{Population}
...
}
}
}
\source{
\url{https://pxdata.stat.fi/PxWeb/pxweb/fi/Postinumeroalueittainen_avoin_tieto/Postinumeroalueittainen_avoin_tieto__uusin/paavo_pxt_12f7.px/}
}
\usage{
vaesto_postinumeroittain
}
\description{
A sample of population data by postal code area from Statistics Finland's Paavo statistics table.
}
\keyword{datasets}
